+++ Library Managment Project/DTOs/LoanBookDTOs/LoanBookRequest.cs
@@ -2,7 +2,7 @@
 {
     public class LoanBookRequest
     {
-        public long MemberCode { get; set; }
+        public int MemberCode { get; set; }
         public int BookCode { get; set; }
     }
 }
